refactor: reorganize page components into dedicated directories
- Move page components into their own directories (address/, block/, tx/, etc.)
- Move Navbar to components/navbar/ directory
- Move devtools components to pages/devtools/
- Fix import organization across multiple files
- Fix accessibility: change anchor to button in Footer
- Remove unused common component files
