+++ src/components/navbar/VersionWarningIcon.tsx
