+++ src/components/pages/tx/TransactionReceiptDisplay.tsx
@@ -1,5 +1,5 @@
 import type React from "react";
-import type { TransactionReceipt } from "../../types";
+import type { TransactionReceipt } from "../../../types";
 
 interface TransactionReceiptDisplayProps {
   receipt: TransactionReceipt;
